Setup docstrings and rename module to giesomat. Allows stand-alone usage.
