+++ lib/python3/green-environment/__init__.py
@@ -1,1 +1,1 @@
-from sensor import GiesOMat
+from giesomat import GiesOMat
